+++ nip/parameters/protocol.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Optional
 
 from nip.parameters.parameters_base import SubParameters, register_parameter_class
-from nip.parameters.types import GuessType
+from nip.parameters.types import GuessType, VerifierDecisionScaleType
 
 
 @register_parameter_class
@@ -27,6 +27,15 @@ class CommonProtocolParameters(SubParameters):
         The reward given to the verifier when it guesses correctly.
     verifier_incorrect_penalty : float
         The penalty given to the verifier when it guesses incorrectly.
+    verifier_neither_accept_nor_reject_reward : float | None
+        The reward given to the verifier when it neither accepts nor rejects. If
+        ``None``, the mid-point between ``verifier_reward`` and
+        ``verifier_incorrect_penalty`` is used. This value is only relevant for
+        text-based scenarios. Note that when using a verifier decision scale (see
+        ``verifier_decision_scale``), reward for intermediate decisions is computed by
+        interpolating piece-wise linearly between ``verifier_incorrect_penalty``,
+        ``verifier_neither_accept_nor_reject_reward`` and ``verifier_reward``. So in
+        this case you probably want to set this to ``None``.
     verifier_terminated_penalty : float
         The reward given to the verifier if the episode terminates before it guesses.
     verifier_no_guess_reward : float
@@ -39,6 +48,10 @@ class CommonProtocolParameters(SubParameters):
         guess using its policy.
     zero_knowledge: bool
         Whether to use a zero-knowledge version of the protocol.
+    verifier_decision_scale : VerifierDecisionScaleType
+        The scale used by the verifier to make its decision. This allows for more
+        nuanced decisions than just "accept" or "reject". This is only relevant for
+        text-based scenarios.
     """
 
     verifier_first: bool = True
@@ -47,6 +60,7 @@ class CommonProtocolParameters(SubParameters):
     prover_reward: float = 1.0
     verifier_reward: float = 1.0
     verifier_incorrect_penalty: float = -1.0
+    verifier_neither_accept_nor_reject_reward: float | None = None
     verifier_terminated_penalty: float = -1.0
     verifier_no_guess_reward: float = 0.0
     shared_reward: bool = False
@@ -54,6 +68,8 @@ class CommonProtocolParameters(SubParameters):
     force_guess: Optional[GuessType] = None
 
     zero_knowledge: bool = False
+
+    verifier_decision_scale: VerifierDecisionScaleType = "accept_reject"
 
 
 @dataclass
@@ -109,7 +125,7 @@ class DebateProtocolParameters(LongProtocolParameters):
         When the provers send messages sequentially, whether prover 0 goes first.
     randomize_channel_order : bool
         Whether to randomize the order of the channels when prompting the verifier. Only
-        relevant in text-based protocols.
+        relevant in text-based scenarios.
     """
 
     sequential: bool = False
@@ -137,7 +153,7 @@ class MnipProtocolParameters(LongProtocolParameters):
         When the provers send messages sequentially, whether prover 0 goes first.
     randomize_channel_order : bool
         Whether to randomize the order of the channels when prompting the verifier. Only
-        relevant in text-based protocols.
+        relevant in text-based scenarios.
     """
 
     sequential: bool = False