+++ nip/utils/plotting/__init__.py
@@ -1,3 +1,5 @@
 """Utilities for plotting the results of NIP experiments."""
 
 from .decision_agreement import DecisionAgreementAnalysis, analyse_decision_agreement
+from .rollouts import get_last_timestep_mask
+from .decision_scale import get_decision_histogram